refactor(backend): extract HostedSiteEventLog from OpenBotDatabase
Moves both hosted-site aggregates — `hosted-site-operation` for a running
operation and `hosted-site-terminal` for a terminal event whose conversation
marker has not been written yet — into
`src/backend/database/hosted-site-event-log.ts`, together with the validators
that refuse a payload the marker could not be rebuilt from.

`PendingHostedSiteTerminalEvent` and `ActiveHostedSiteConversationEvent` move
with the controller and are re-exported from the facade: both were declared in
this module before the split and are part of its frozen public surface.

The controller opens no transaction of its own; it dispatches through the core,
so a caller already inside one has these writes pulled into it.
